Scorpion/requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
pillow==11.2.1
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
validators==0.35.0
